back-end/ai_service/requirements.txt
fastapi
uvicorn
openai
faiss-cpu
numpy
python-dotenv
python-json-logger
pytest
httpx
pytest-asyncio